feat: refactor request and response example
